backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.2,<0.116
uvicorn[standard]>=0.27.1

# Database
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.2

# Celery & Redis
celery[redis]>=5.3.6
redis>=5.0.1

# Validation & Utils (use versions with pre-built wheels for Python 3.13)
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
email-validator>=2.1.0

# CORS
python-multipart>=0.0.9
